client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "grpcio-tools>=1.70.0",
    "grpcio>=1.70.0",
    "opencv-python>=4.11.0.86",
]
